// Frontend/src/app/shared/model/sucurales-afectadas.ts
export interface SucursalesAfectadas {
  idCard: number;
  title: string;
  alcanzabilidad: number;
  disponibilidad: number;
  perdidaPaquetes: number;
  totalAfectaciones?: number;
  tiempoCaidaDescripcion?: string;
  progressChartdata?: {};
  classCard?: string;
  classIcon?: string;
  total?: number;
  tiempoCaida?: number;
  idDispositivo?: string;
  ipNs?:string;
}